langchain-azure/pyproject.toml
[project]
name = "spyglass-example-langchain-azure"
version = "0.1.0"
description = "LangChain Azure OpenAI example with Spyglass tracing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.1.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "spyglass-ai",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["langchain-azure.py"]

# Uncomment the section below to use the local spyglass-sdk for development
# This assumes spyglass-sdk is located at ../../spyglass-sdk relative to this file
[tool.uv.sources]
spyglass-ai = { path = "../../spyglass-sdk", editable = true }
